// src/app/model/CardPower.ts
export enum CardPower {
    takeCandy = "TAKE_CANDY",
    drawOne = "DRAW_ONE",
    drawTwo = "DRAW_TWO",
    start = "START",
    drinkCofee = "DRINK_COFEE",
    belowTheStack = "BELOW_THE_STACK",
    cardsIntoPast = "CARDS_INTO_PAST",
    cardsFromPast = "CARDS_FROM_PAST",
    cardIntoFuture = "CARD_INTO_FUTURE",
    allCardsIntoFuture = "ALL_CARDS_INTO_FUTURE",
    exchangeCard = "EXCHANGE_CARD",
    resetCandies = "RESET_CANDIES"
}